fix(dialyzer): also use ERTS version
If we update the point release of Erlang, we keep the same major version. However, the Dialyzer PLT cache is invalid and not detected.

This incorporates the ERTS (Erlang Run-Time System) version to better detect these cases.

For an example of this invalid cache, see https://github.com/mbta/skate/runs/4969819997?check_suite_focus=true