FEAT: new ecs-run-task action for starting EventBridge tasks via GHA (#62)
The change adds a new action that allows an ECS task, that is usually run via EventBridge scheduling, to be manually started via Github Action.

The action works by accepting a CLUSTER and SERVICE env var. Using the defined CLUSTER and SERVICE, information is pulled from the aws ecs describe-services command and used to trigger an aws ecs run-task command. The latest task definition, for the defined SERVICE will be triggered via run-task.