chore: update `dialyzer` action dependencies
This action depended on an outdated version of `@actions/cache` which
will stop working entirely on April 15th, 2025, as the underlying cache
service is being shut down[1].

Additionally, the action runs on the `node16` image which is no longer
supported by GitHub[2] as of March 2024 (in addition to Node 16 itself
being unsupported since September 2023).

[1]: https://github.blog/changelog/2025-03-20-notification-of-upcoming-breaking-changes-in-github-actions/
[2]: https://github.blog/changelog/2024-03-07-github-actions-all-actions-will-run-on-node20-instead-of-node16-by-default/

Since this action has few dependencies and few changes are required to
update all of them, this commit does just that, rather than attempt to
work out a minimal change that *only* addresses the above issues.

Notable changes beyond version updates:

* With updated `typescript` and `@vercel/ncc`, the build process
  encounters several new type errors due to missing NodeJS types.
  `@types/node` is added to address this. `@tsconfig/node20` is also
  added and used in `tsconfig.json` to ensure our TypeScript settings
  are aligned with the environment we're targeting.

* `@actions/glob` now exposes a `hashFiles` utility, so we don't need
  to include our own implementation.